Move AuthDTO out of AccountModels
